==> Single nucleus omics/05_Gene regulatory networks/00_Construct ArchR metadata.ipynb <==
### Load needed libraries

import os
import warnings
import pandas as pd
warnings.filterwarnings("ignore")

pwd = os.getcwd()

### Load needed datasets

# From https://sea-ad-single-cell-profiling.s3.amazonaws.com/index.html#MTG/RNAseq/Supplementary%20Information/
metadata = pd.read_csv(os.path.join(pwd, "input", "SEAAD_MTG_ATACseq_final-nuclei_metadata.2024-02-13.csv"), index_col=0)
metadata = metadata.loc[(metadata["Class"] == "Non-neuronal and Non-neural") & (metadata["Severely Affected Donor"] == "N"), :].copy()

# Internal list of pre-built Arrow files
inputf = pd.read_csv(os.path.join(pwd, "input", "ATAC_AD_Center_Grant_update.csv"), index_col=0)


### Add back Arrow file cell IDs

inputf = inputf.loc[:, ["library_prep", "arrow_file"]]
inputf["arrow"] = inputf['arrow_file'].str.replace(".arrow", "")

metadata = metadata.merge(inputf, how="left", left_on="library_prep", right_on="library_prep")

metadata["archr_index"] = metadata["arrow"] + "#" + metadata["bc"] + '-1'
metadata = metadata.loc[~metadata["archr_index"].isnull(), :]

### Save metadata

metadata.to_csv(os.path.join(pwd, "input", "Non-neuronal_Supertype_Annotation_with_archr_index.csv"), index=False)
